minor refactoring of the classes
